SalaDBot-main/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
python-dotenv==1.0.1
supabase==2.10.0
openai==1.55.3
requests==2.32.3
pydantic==2.10.3
pytz
